Update _base.py
Fixed an issue where the 'statusUrl' property in the response from a create replica command on a feature service results in a false-positive test on the line that's checking for 'status' in a dictionary object.  The result is an error on the next line that attempts to access 'status' as a named dictionary property, when the results parameter is actually just a string.  I don't know if this is the best fix for this...