Update _clusters.py
Changed default value of tcpClusterPort argument in createCluster() (from `None` to `""`) . Because when creating a cluster without specifying the port and having None as the value, `500 [u"tcpClusterPort value 'None' is invalid. tcpClusterPort should be a positive integer."]` error message is displayed. After changing to `""`, a suitable integer value is being used.